Refactor SmeupInterpreterTest and moved T60 unit tests in MULANGT60VideoTest
