+++ rpgJavaInterpreter-core/src/test/kotlin/com/smeup/rpgparser/evaluation/SmeupInterpreterTest.kt
@@ -468,12 +468,6 @@ open class SmeupInterpreterTest : AbstractTest() {
     }
 
     @Test
-    fun executeT60_A10_P01_02() {
-        val expected = listOf<String>()
-        assertEquals(expected, "smeup/T60_A10_P01-02".outputOf(configuration = smeupConfig))
-    }
-
-    @Test
     fun executeT10_A20_P35_38() {
         val expected = listOf<String>(
             "Res(21, 0, 0, 0, 0, 0)",
@@ -551,12 +545,6 @@ open class SmeupInterpreterTest : AbstractTest() {
             "A50_A91(Funzione  ) A50_N91(Funzione  )A50_V91(Funzione)"
         )
         assertEquals(expected, "smeup/T02_A50_P09".outputOf())
-    }
-
-    @Test
-    fun executeT60_A10_P01() {
-        val expected = listOf<String>("KA(0)KB(0)KC(0)KD(0)KE(0)KF(0)KG(0)KH(0)KI(0)KJ(0)KK(0)KL(0)KM(0)KN(0)KP(0)KQ(0)KR(0)KS(0)KT(0)KU(0)KV(0)KW(0)KX(0)KY(0)")
-        assertEquals(expected, "smeup/T60_A10_P01".outputOf(configuration = smeupConfig))
     }
 
     @Test
